Moved HoverPopupWindow methods to a dedicated class: PopupSelect
